+++ client/src/landing/NavBar.js
@@ -23,51 +23,23 @@
  *  NavBar for logged-in and logged-out users.
  */
 
-import { faGitlab } from "@fortawesome/free-brands-svg-icons";
-import { faQuestionCircle, faUser } from "@fortawesome/free-regular-svg-icons";
-import { faBars, faPlus } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { Component, Fragment } from "react";
-import { Search } from "react-bootstrap-icons";
-import { Link, Route, Switch } from "react-router-dom";
-import {
-  Collapse,
-  DropdownItem,
-  DropdownMenu,
-  DropdownToggle,
-  Nav,
-  NavItem,
-  Navbar,
-  NavbarToggler,
-  UncontrolledDropdown,
-} from "reactstrap";
-
 import { useSelector } from "react-redux";
-import { LoginHelper } from "../authentication";
-import { UserAvatar } from "../components/Avatar";
-import { ExternalDocsLink, ExternalLink } from "../components/ExternalLinks";
-import { Loader } from "../components/Loader";
+import { Link, Route, Switch } from "react-router-dom";
+import { Nav, Navbar } from "reactstrap";
+
+import { ExternalDocsLink } from "../components/ExternalLinks";
 import { RenkuNavLink } from "../components/RenkuNavLink";
-import { parseChartVersion } from "../help/HelpRelease";
-import { NotificationsMenu } from "../notifications";
-import { StatuspageBanner } from "../statuspage";
-import { Docs, Links, RenkuPythonDocs } from "../utils/constants/Docs";
-import {
-  getActiveProjectPathWithNamespace,
-  gitLabUrlFromProfileUrl,
-} from "../utils/helpers/HelperFunctions";
-import { Url } from "../utils/helpers/url";
-import AdminDropdownItem from "./AdminDropdownItem";
-import { NavBarWarnings } from "./NavBarWarnings";
+import AnonymousNavBar2 from "../components/navbar/AnonymousNavBar";
 import LoggedInNavBar2 from "../components/navbar/LoggedInNavBar";
+import { RENKU_LOGO } from "../components/navbar/navbar.constans";
+import { parseChartVersion } from "../help/HelpRelease";
+import { Links } from "../utils/constants/Docs";
+import { Url } from "../utils/helpers/url";
 
 import "./NavBar.css";
 
-const logo = "/static/public/img/logo.svg";
-
 function RenkuNavBar(props) {
   const { user } = props;
-  const userAvatar = <UserAvatar person={user.data} size="sm" />;
   const projectMetadata = useSelector(
     (state) => state.stateModel.project?.metadata
   );
@@ -82,411 +54,21 @@ function RenkuNavBar(props) {
       <Route path={sessionShowUrl} />
       <Route>
         {user.logged ? (
-          <LoggedInNavBar2 {...props} userAvatar={userAvatar} />
+          <LoggedInNavBar2
+            model={props.model}
+            notifications={props.notifications}
+            params={props.params}
+          />
         ) : (
-          <AnonymousNavBar {...props} userAvatar={userAvatar} />
+          <AnonymousNavBar2
+            model={props.model}
+            notifications={props.notifications}
+            params={props.params}
+          />
         )}
       </Route>
     </Switch>
   );
-}
-
-function gitLabSettingsUrlFromProfileUrl(webUrl) {
-  // Yes, the settings URL ends with 'profile'; the profile URL ends with the username
-  const comps = webUrl.split("/");
-  comps.pop();
-  comps.push("-/profile");
-  return comps.join("/");
-}
-
-class RenkuToolbarItemUser extends Component {
-  render() {
-    const { location, user } = this.props;
-    const gatewayURL = this.props.params.GATEWAY_URL;
-    const uiserverURL = this.props.params.UISERVER_URL;
-    const redirect_url = encodeURIComponent(this.props.params.BASE_URL);
-    if (!user.fetched) {
-      return <Loader inline size={16} />;
-    } else if (!user.logged) {
-      const to = Url.get(Url.pages.login.link, { pathname: location.pathname });
-      return <RenkuNavLink to={to} title="Login" />;
-    }
-
-    return (
-      <UncontrolledDropdown className="nav-item dropdown">
-        <DropdownToggle className="nav-link" nav caret id="profile-dropdown">
-          <FontAwesomeIcon icon={faUser} id="userIcon" />
-        </DropdownToggle>
-        <DropdownMenu
-          className="user-menu btn-with-menu-options"
-          end
-          key="user-bar"
-          aria-labelledby="user-menu"
-        >
-          <DropdownItem className="p-0">
-            <ExternalLink
-              url={`${gatewayURL}/auth/user-profile`}
-              title="Account"
-              className="dropdown-item"
-              role="link"
-            />
-          </DropdownItem>
-          <AdminDropdownItem />
-          <DropdownItem divider />
-          <a
-            id="logout-link"
-            className="dropdown-item"
-            onClick={() => {
-              LoginHelper.notifyLogout();
-            }}
-            href={`${uiserverURL}/auth/logout?redirect_url=${redirect_url}`}
-          >
-            Logout
-          </a>
-        </DropdownMenu>
-      </UncontrolledDropdown>
-    );
-  }
-}
-
-class RenkuToolbarItemPlus extends Component {
-  render() {
-    // Display the Issue/Notebook server related header options only if a project is active.
-    const activeProjectPathWithNamespace = getActiveProjectPathWithNamespace(
-      this.props.currentPath
-    );
-    const datasetDropdown = activeProjectPathWithNamespace ? (
-      <DropdownItem className="p-0">
-        <Link
-          className="dropdown-item"
-          id="navbar-dataset-new"
-          to={`/projects/${activeProjectPathWithNamespace}/datasets/new`}
-        >
-          Dataset
-        </Link>
-      </DropdownItem>
-    ) : null;
-    const projectDropdown = (
-      <DropdownItem className="p-0">
-        <Link
-          className="dropdown-item"
-          id="navbar-project-new"
-          to="/projects/new"
-        >
-          Project
-        </Link>
-      </DropdownItem>
-    );
-
-    return (
-      <UncontrolledDropdown className="nav-item dropdown">
-        <Fragment>
-          <DropdownToggle className="nav-link" nav caret id="plus-dropdown">
-            <FontAwesomeIcon icon={faPlus} id="createPlus" />
-          </DropdownToggle>
-          <DropdownMenu
-            className="plus-menu btn-with-menu-options"
-            end
-            key="plus-bar"
-            aria-labelledby="plus-menu"
-          >
-            {projectDropdown}
-            {datasetDropdown}
-          </DropdownMenu>
-        </Fragment>
-      </UncontrolledDropdown>
-    );
-  }
-}
-
-function RenkuToolbarHelpMenu() {
-  return (
-    <UncontrolledDropdown className="nav-item dropdown">
-      <Fragment>
-        <DropdownToggle className="nav-link" nav caret>
-          <FontAwesomeIcon icon={faQuestionCircle} id="helpDropdownToggle" />
-        </DropdownToggle>
-        <DropdownMenu
-          className="help-menu btn-with-menu-options"
-          key="help-bar"
-          aria-labelledby="help-menu"
-        >
-          <DropdownItem className="p-0">
-            <Link className="dropdown-item" to="/help">
-              Help
-            </Link>
-          </DropdownItem>
-          <DropdownItem divider />
-          <DropdownItem className="p-0">
-            <ExternalDocsLink
-              url={Docs.READ_THE_DOCS_ROOT}
-              title="Renku Docs"
-              className="dropdown-item"
-            />
-          </DropdownItem>
-          <DropdownItem className="p-0">
-            <ExternalDocsLink
-              url={RenkuPythonDocs.READ_THE_DOCS_ROOT}
-              title="Renku CLI Docs"
-              className="dropdown-item"
-            />
-          </DropdownItem>
-          <DropdownItem divider />
-          <DropdownItem className="p-0">
-            <ExternalDocsLink
-              url={Links.DISCOURSE}
-              title="Forum"
-              className="dropdown-item"
-            />
-          </DropdownItem>
-          <DropdownItem className="p-0">
-            <ExternalDocsLink
-              url={Links.GITTER}
-              title="Gitter"
-              className="dropdown-item"
-            />
-          </DropdownItem>
-          <DropdownItem className="p-0">
-            <ExternalDocsLink
-              url={Links.GITHUB}
-              title="GitHub"
-              className="dropdown-item"
-            />
-          </DropdownItem>
-        </DropdownMenu>
-      </Fragment>
-    </UncontrolledDropdown>
-  );
-}
-
-function RenkuToolbarGitLabMenu(props) {
-  const user = props.user;
-  if (!user.fetched) return "";
-  else if (!user.data.id) return "";
-
-  const gitLabUrl = gitLabUrlFromProfileUrl(user.data.web_url);
-
-  return (
-    <UncontrolledDropdown className="nav-item dropdown">
-      <Fragment>
-        <DropdownToggle className="nav-link" nav caret id="gitLab-menu">
-          <FontAwesomeIcon icon={faGitlab} id="gitLabDropdownToggle" />
-        </DropdownToggle>
-        <DropdownMenu
-          className="gitLab-menu btn-with-menu-options"
-          end
-          key="gitLab-bar"
-          aria-labelledby="gitLab-menu"
-        >
-          <DropdownItem className="p-0">
-            <ExternalLink
-              url={gitLabUrl}
-              title="GitLab"
-              className="dropdown-item"
-              role="link"
-            />
-          </DropdownItem>
-          <DropdownItem className="p-0">
-            <ExternalLink
-              url={gitLabSettingsUrlFromProfileUrl(user.data.web_url)}
-              title="Settings"
-              className="dropdown-item"
-              role="link"
-            />
-          </DropdownItem>
-          <DropdownItem className="p-0">
-            <ExternalLink
-              url={user.data.web_url}
-              title="Profile"
-              className="dropdown-item"
-              role="link"
-            />
-          </DropdownItem>
-        </DropdownMenu>
-      </Fragment>
-    </UncontrolledDropdown>
-  );
-}
-
-function RenkuToolbarNotifications(props) {
-  if (!props.notifications) return null;
-
-  return (
-    <UncontrolledDropdown className="nav-item dropdown">
-      <NotificationsMenu {...props} />
-    </UncontrolledDropdown>
-  );
-}
-
-class LoggedInNavBar extends Component {
-  constructor(props) {
-    super(props);
-    this.onSelect = this.handleSelect.bind(this);
-    this.toggle = this.toggle.bind(this);
-    this.state = {
-      isOpen: true,
-    };
-  }
-
-  toggle() {
-    this.setState({
-      isOpen: !this.state.isOpen,
-    });
-  }
-
-  handleSelect(eventKey) {
-    let nextRoute = null;
-    switch (eventKey) {
-      case "new.projects":
-        nextRoute = "/projects";
-        break;
-      default:
-        break;
-    }
-    if (null != nextRoute) this.props.history.push(nextRoute);
-  }
-  render() {
-    const searchIcon = <Search />;
-    return (
-      <Fragment>
-        <header className="navbar navbar-expand-lg navbar-dark rk-navbar p-0">
-          <Navbar
-            color="primary"
-            className="container-fluid flex-wrap flex-lg-nowrap renku-container"
-          >
-            <Link
-              id="link-home"
-              data-cy="link-home"
-              to="/"
-              className="navbar-brand me-2 pb-0 pt-0"
-            >
-              <img src={logo} alt="Renku" height="50" className="d-block" />
-            </Link>
-            <NavbarToggler onClick={this.toggle} className="border-0">
-              <FontAwesomeIcon icon={faBars} id="userIcon" color="white" />
-            </NavbarToggler>
-            <Collapse isOpen={!this.state.isOpen} navbar className="">
-              <Nav className="navbar-nav flex-row flex-nowrap ms-lg-auto">
-                <NavItem className="nav-item col-4 col-lg-auto pe-lg-4">
-                  <RenkuNavLink
-                    to="/search"
-                    title="Search"
-                    id="link-search"
-                    icon={searchIcon}
-                    className="d-flex gap-2 align-items-center"
-                  />
-                </NavItem>
-                <NavItem
-                  id="link-dashboard"
-                  data-cy="link-dashboard"
-                  to="/"
-                  className="nav-item col-4 col-lg-auto pe-lg-4"
-                >
-                  <RenkuNavLink to="/" title="Dashboard" id="link-dashboard" />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarItemPlus
-                    currentPath={this.props.location.pathname}
-                  />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarGitLabMenu user={this.props.user} />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarHelpMenu />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarNotifications {...this.props} />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarItemUser {...this.props} />
-                </NavItem>
-              </Nav>
-            </Collapse>
-          </Navbar>
-        </header>
-        <StatuspageBanner
-          siteStatusUrl={Url.get(Url.pages.help.status)}
-          model={this.props.model}
-          location={this.props.location}
-        />
-        <NavBarWarnings
-          model={this.props.model}
-          uiShortSha={this.props.params["UI_SHORT_SHA"]}
-        />
-      </Fragment>
-    );
-  }
-}
-
-class AnonymousNavBar extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isOpen: true,
-    };
-    this.toggle = this.toggle.bind(this);
-  }
-
-  toggle() {
-    this.setState({
-      isOpen: !this.state.isOpen,
-    });
-  }
-
-  render() {
-    const searchIcon = <Search />;
-    return (
-      <Fragment>
-        <header className="navbar navbar-expand-lg navbar-dark rk-navbar p-0">
-          <Navbar
-            color="primary"
-            className="container-fluid flex-wrap flex-lg-nowrap renku-container"
-          >
-            <Link id="link-home" to="/" className="navbar-brand me-2 pb-0 pt-0">
-              <img src={logo} alt="Renku" height="50" className="d-block" />
-            </Link>
-            <NavbarToggler onClick={this.toggle} className="border-0">
-              <FontAwesomeIcon icon={faBars} id="userIcon" color="white" />
-            </NavbarToggler>
-            <Collapse isOpen={!this.state.isOpen} navbar className="">
-              <Nav className="navbar-nav flex-row flex-nowrap ms-lg-auto">
-                <NavItem className="nav-item col-4 col-lg-auto pe-lg-4">
-                  <RenkuNavLink
-                    to="/search"
-                    title="Search"
-                    id="link-search"
-                    icon={searchIcon}
-                    className="d-flex gap-2 align-items-center"
-                  />
-                </NavItem>
-                <NavItem className="nav-item col-4 col-lg-auto pe-4">
-                  <RenkuNavLink
-                    to="/sessions"
-                    title="Sessions"
-                    id="link-sessions"
-                  />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarHelpMenu />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarNotifications {...this.props} />
-                </NavItem>
-                <NavItem className="nav-item col-1 col-lg-auto">
-                  <RenkuToolbarItemUser {...this.props} />
-                </NavItem>
-              </Nav>
-            </Collapse>
-          </Navbar>
-        </header>
-        <NavBarWarnings
-          model={this.props.model}
-          uiShortSha={this.props.params["UI_SHORT_SHA"]}
-        />
-      </Fragment>
-    );
-  }
 }
 
 function FooterNavbarAnonymousLinks() {
@@ -568,7 +150,7 @@ function FooterNavbar({ location, params }) {
             data-cy="version-info"
           >
             <Link className="nav-link" to={Url.pages.help.release}>
-              <img src={logo} alt="Renku" className="pb-2" height="44" />
+              <img src={RENKU_LOGO} alt="Renku" className="pb-2" height="44" />
               <span className="ps-2">{displayVersion}</span>
             </Link>
           </Nav>
@@ -595,9 +177,4 @@ function FooterNavbar({ location, params }) {
   );
 }
 
-export {
-  FooterNavbar,
-  RenkuNavBar,
-  RenkuToolbarHelpMenu,
-  RenkuToolbarNotifications,
-};
+export { FooterNavbar, RenkuNavBar };
