+++ client/src/landing/anonymousHomeNav.tsx
@@ -16,20 +16,15 @@
  * limitations under the License.
  */
 
-import React from "react";
 import type { CSSProperties } from "react";
-import { Button, Col, Collapse, Nav, Navbar, NavItem, Row } from "reactstrap";
+import React from "react";
+import { List } from "react-bootstrap-icons";
 import { Link } from "react-router-dom";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBars } from "@fortawesome/free-solid-svg-icons";
-
+import { Button, Col, Collapse, Nav, Navbar, NavItem, Row } from "reactstrap";
 import { ExternalLink } from "../components/ExternalLinks";
-import { Url } from "../utils/helpers/url";
 import { Docs, Links, RenkuPythonDocs } from "../utils/constants/Docs";
-
+import { Url } from "../utils/helpers/url";
 import type { AnonymousHomeConfig } from "./anonymousHome.types";
-import { RenkuToolbarNotifications } from "./NavBar";
 
 const logo = "/static/public/img/logo.svg";
 
@@ -156,7 +151,7 @@ function TopNavLink({ title, to }: BottomNavLinkProps) {
   );
 }
 
-function TopNav(props: AnonymousHomeConfig) {
+function TopNav() {
   const [isOpen, setIsOpen] = React.useState(false);
   const toggleOpen = () => setIsOpen(!isOpen);
   return (
@@ -184,7 +179,7 @@ function TopNav(props: AnonymousHomeConfig) {
             Login
           </Link>
           <Button onClick={toggleOpen} id="nav-hamburger" className="border-0">
-            <FontAwesomeIcon className="m-0" icon={faBars} id="userIcon" />
+            <List className="m-0 bi" />
           </Button>
         </div>
       </header>
@@ -224,9 +219,6 @@ function TopNav(props: AnonymousHomeConfig) {
               </NavItem>
               <NavItem className="d-block d-md-none nav-item">
                 <TopNavLink title="Help" to={Url.get(Url.pages.help)} />
-              </NavItem>
-              <NavItem className="nav-item mb-2">
-                <RenkuToolbarNotifications {...props} />
               </NavItem>
             </Nav>
           </Navbar>
