+++ client/src/notifications/Notifications.present.js
@@ -23,35 +23,34 @@
  *  Presentational components for notifications
  */
 
+import {
+  faCheck,
+  faCheckCircle,
+  faExclamationTriangle,
+  faInfoCircle,
+  faLink,
+  faTimes,
+} from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Component, Fragment, useState } from "react";
+import { InboxFill } from "react-bootstrap-icons";
 import { Link } from "react-router-dom";
 import {
   Badge,
-  DropdownMenu,
-  DropdownToggle,
-  DropdownItem,
   Button,
-  Row,
+  CardBody,
   Col,
   Collapse,
-  CardBody,
+  DropdownItem,
+  DropdownMenu,
+  DropdownToggle,
+  Row,
 } from "reactstrap";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faLink,
-  faInfoCircle,
-  faExclamationTriangle,
-  faInbox,
-  faTimes,
-  faCheck,
-  faCheckCircle,
-} from "@fortawesome/free-solid-svg-icons";
-
 import { NotificationsInfo } from ".";
-
-import "./Notifications.css";
 import { ExternalLink } from "../components/ExternalLinks";
 import { TimeCaption } from "../components/TimeCaption";
+
+import "./Notifications.css";
 
 /**
  * Close button for the toast notification.
@@ -276,7 +275,7 @@ class NotificationsMenu extends Component {
       <Fragment>
         {/* This throws an error in test: Warning `Reference` should not be used outside of a `Manager` component. */}
         <DropdownToggle className="nav-link" nav caret>
-          <FontAwesomeIcon icon={faInbox} id="notificationsBarIcon" />
+          <InboxFill className="bi" id="notificationsBarIcon" />
           {badge}
         </DropdownToggle>
         <DropdownMenu
@@ -557,10 +556,10 @@ function NotificationPageItemDetails(props) {
 }
 
 export {
-  NotificationsMenu,
-  NotificationToast,
-  NotificationDropdownItem,
   CloseToast,
-  Notifications,
+  NotificationDropdownItem,
   NotificationPageItem,
+  NotificationToast,
+  Notifications,
+  NotificationsMenu,
 };
